# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Current": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          },
          "hash": {
            "oneOf": [
              {
                "pattern": "^sha256-[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "enum": ["put", "delete", "rollback"],
            "type": "string"
          },
          "version": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["version", "hash", "deleted", "kind", "author", "createdAt"],
        "type": "object"
      },
      "DeleteResult": {
        "additionalProperties": false,
        "properties": {
          "changeId": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "version": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["version", "changeId", "createdAt"],
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "validation",
              "invalid-path",
              "body-not-well-formed",
              "unauthorized",
              "scope",
              "not-found",
              "file-deleted",
              "version-not-found",
              "stale",
              "exists",
              "already-deleted",
              "already-rotated",
              "token-expired",
              "rate-limited",
              "payload-too-large",
              "idempotency-key-reused",
              "rollback-target-tombstone",
              "internal"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "successorId": {
            "type": "string"
          }
        },
        "required": ["code", "message"],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "current": {
            "$ref": "#/components/schemas/Current"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": ["error"],
        "type": "object"
      },
...
      "FileSummary": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "hash": {
            "oneOf": [
              {
                "pattern": "^sha256-[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "headVersion": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "path": {
            "type": "string"
          },
          "size": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          }
        },
        "required": ["path", "headVersion", "hash", "size", "deleted", "updatedAt"],
        "type": "object"
      },
...
      "ListFilesResult": {
        "additionalProperties": false,
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileSummary"
            },
            "type": "array"
          },
          "nextAfter": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["files", "nextAfter"],
        "type": "object"
      },
...
      "RollbackResult": {
        "additionalProperties": false,
        "properties": {
          "changeId": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "hash": {
            "pattern": "^sha256-[0-9a-f]{64}$",
            "type": "string"
          },
          "identicalToHead": {
            "type": "boolean"
          },
          "rollbackOf": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "version": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": ["version", "hash", "rollbackOf", "identicalToHead", "changeId", "createdAt"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Credentials for one stash cannot reveal another stash: foreign stashes are concealed as 404 responses.\n\nLists and change feeds use keyset cursors (after, before, or since); limit defaults to 50 and has a maximum of 200.\n\nFile reads use ETags, If-None-Match, and 304 responses for conditional requests.\n\nMutation idempotency keys are retained in a 7-day ledger; reusing a key for a different request returns 422 idempotency-key-reused.",
    "title": "zudo-history-stash",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/stashes/{stash}/delete/{path}": {
      "post": {
        "description": "Appends a tombstone version for a file using compare-and-set semantics.\n\nwrite; administrator or a matching write stash token.\n\nThe path value contains unescaped `/`. OpenAPI 3.1 path templating does not permit this, so generated clients must not be assumed to work for this operation.",
        "operationId": "deleteFile",
        "parameters": [
          {
            "description": "A 1-63 character stash name containing lowercase ASCII letters, digits, or hyphens and beginning with a letter or digit.",
            "in": "path",
            "name": "stash",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "A non-empty slash-separated path whose segments contain only ASCII letters, digits, dots, underscores, or hyphens and are neither dot nor dot-dot; maximum 512 UTF-8 bytes. The path value contains unescaped `/`. OpenAPI 3.1 path templating does not permit this, so generated clients must not be assumed to work for this operation.",
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-wildcard": true
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "author": {
                    "description": "Must be a well-formed Unicode string. Maximum 200 UTF-8 bytes.",
                    "type": "string"
                  },
                  "expectedVersion": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  },
                  "message": {
                    "description": "Must be a well-formed Unicode string. Maximum 2000 UTF-8 bytes.",
                    "type": "string"
                  }
                },
                "required": ["expectedVersion"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "changeId": 8,
                  "createdAt": "2026-08-26T01:00:00.000Z",
                  "version": 4
                },
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            },
            "description": "A tombstone version was appended.",
            "headers": {
              "Idempotent-Replayed": {
                "description": "Whether the idempotency ledger replayed the response.",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `validation`, `invalid-path`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `scope`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `not-found`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `stale` (includes current), `already-deleted` (includes current)."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `payload-too-large`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `idempotency-key-reused`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `rate-limited`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying a rate-limited request.",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `internal`."
          }
        },
        "summary": "Delete a file",
        "x-principal": "write",
        "x-wildcard": true
      }
    },
...
    "/v1/stashes/{stash}/files": {
      "get": {
        "description": "Returns a keyset-paginated list of file summaries.\n\nread; administrator or a matching read/write stash token.",
        "operationId": "listFiles",
        "parameters": [
          {
            "description": "A 1-63 character stash name containing lowercase ASCII letters, digits, or hyphens and beginning with a letter or digit.",
            "in": "path",
            "name": "stash",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "includeDeleted",
            "required": false,
            "schema": {
              "default": false,
              "type": "boolean"
            },
            "style": "form"
          },
          {
            "explode": false,
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            },
            "style": "form"
          },
          {
            "explode": false,
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "files": [
                    {
                      "deleted": false,
                      "hash": "sha256-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "headVersion": 3,
                      "path": "docs/guide.md",
                      "size": 13,
                      "updatedAt": "2026-08-26T01:00:00.000Z"
                    }
                  ],
                  "nextAfter": null
                },
                "schema": {
                  "$ref": "#/components/schemas/ListFilesResult"
                }
              }
            },
            "description": "A page of file summaries."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `validation`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `unauthorized`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `not-found`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `rate-limited`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying a rate-limited request.",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "List files",
        "x-principal": "read"
      }
    },
...
    "/v1/stashes/{stash}/rollback/{path}": {
      "post": {
        "description": "Restores a target version by appending a new rollback version.\n\nwrite; administrator or a matching write stash token.\n\nThe path value contains unescaped `/`. OpenAPI 3.1 path templating does not permit this, so generated clients must not be assumed to work for this operation.",
        "operationId": "rollbackFile",
        "parameters": [
          {
            "description": "A 1-63 character stash name containing lowercase ASCII letters, digits, or hyphens and beginning with a letter or digit.",
            "in": "path",
            "name": "stash",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "A non-empty slash-separated path whose segments contain only ASCII letters, digits, dots, underscores, or hyphens and are neither dot nor dot-dot; maximum 512 UTF-8 bytes. The path value contains unescaped `/`. OpenAPI 3.1 path templating does not permit this, so generated clients must not be assumed to work for this operation.",
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-wildcard": true
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "author": {
                    "description": "Must be a well-formed Unicode string. Maximum 200 UTF-8 bytes.",
                    "type": "string"
                  },
                  "expectedVersion": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  },
                  "message": {
                    "description": "Must be a well-formed Unicode string. Maximum 2000 UTF-8 bytes.",
                    "type": "string"
                  },
                  "meta": {
                    "additionalProperties": {
                      "description": "Any JSON value. The recursive z.json() validator is intentionally represented as an unconstrained schema."
                    },
                    "description": "Serialized JSON must be at most 4096 UTF-8 bytes.",
                    "propertyNames": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "toVersion": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  }
                },
                "required": ["toVersion", "expectedVersion"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "changeId": 9,
                  "createdAt": "2026-08-26T01:00:00.000Z",
                  "hash": "sha256-aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                  "identicalToHead": false,
                  "rollbackOf": 3,
                  "version": 5
                },
                "schema": {
                  "$ref": "#/components/schemas/RollbackResult"
                }
              }
            },
            "description": "A rollback version was appended.",
            "headers": {
              "Idempotent-Replayed": {
                "description": "Whether the idempotency ledger replayed the response.",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `validation`, `invalid-path`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `unauthorized`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `scope`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `not-found`, `version-not-found` (includes current)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `stale` (includes current)."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `payload-too-large`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `idempotency-key-reused`, `rollback-target-tombstone` (includes current)."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `rate-limited`.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying a rate-limited request.",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Errors: `internal`."
          }
        },
        "summary": "Roll back a file",
        "x-principal": "write",
        "x-wildcard": true
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "https://stash.example.com"
    }
  ]
}